add ASLR test

+++ mts/reallocate-stack.cpp
@@ -1,20 +1,21 @@
 #include "include/gcc_builtin.hpp"
 #include "include/mss.hpp"
+#include <cstdint>
 
 charBuffer *pb;
 
-long long FORCE_NOINLINE helper(bool option) {
+intptr_t FORCE_NOINLINE helper(bool option) {
   charBuffer buffer;
   if(option) {
-    return reinterpret_cast<long long>(buffer.data);
+    return reinterpret_cast<intptr_t>(buffer.data);
   } else {
     pb = &buffer;
-    return reinterpret_cast<long long>(buffer.data);
+    return reinterpret_cast<intptr_t>(buffer.data);
   }
 }
 
 int main() {
-  long long rv0 = helper(false);
-  long long rv1 = helper(true);
+  intptr_t rv0 = helper(false);
+  intptr_t rv1 = helper(true);
   return rv0-rv1;
 }
